Alterando nomes de arquivos dentro de src/components
