+++ src/App.tsx
@@ -6,10 +6,10 @@ import Dashboard from "./pages/Dashboard";
 import EscolasCadastradas from "./pages/EscolasCadastradas";
 import Home from "./pages/Home";
 import Login from "./pages/Login";
-import RecoverPassword from "./pages/RecoverPassword";
-import Register from "./pages/Register";
-import RegisterSchool from "./pages/RegisterSchool";
-import ResetPassword from "./pages/ResetPassword";
+import RecoverPassword from "./pages/RecuperarSenha";
+import Register from "./pages/CadastrarUsuario";
+import RegisterSchool from "./pages/CadastrarEscola";
+import ResetPassword from "./pages/RedefinirSenha";
 import SolicitacaoAcao from "./pages/SolicitacaoAcao";
 import TelaUPS from "./pages/TelaUPS";
 import { AuthContext } from "./provider/Authentication";
