+++ src/components/cadastrarEscolas/Step1.tsx
