+++ src/pages/EscolasCadastradas.tsx
@@ -1,8 +1,8 @@
-import BreadCrumb from "../components/BreadCrumb";
-import FiltragemTabela from "../components/components-escolasCadastradas/FiltragemTabela";
-import Footer from "../components/components-escolasCadastradas/Footer";
-import Header from "../components/components-escolasCadastradas/Header";
-import TabelaEscola from "../components/components-escolasCadastradas/TabelaEscola";
+import BreadCrumb from "../components/escolasCadastradas/BreadCrumb";
+import FiltragemTabela from "../components/escolasCadastradas/FiltragemTabela";
+import Footer from "../components/Footer";
+import Header from "../components/escolasCadastradas/HeaderListaEscolas";
+import TabelaEscola from "../components/escolasCadastradas/TabelaEscola";
 import { FiltroProvider } from "../context/FiltroTabela";
 import "../styles/App.css";
